+++ Runtime/Extensions/MonoBehaviourExtension.cs
@@ -22,13 +22,5 @@ namespace DT.UniStart {
       }
       mb.StartCoroutine(InvokeRepeatingRoutine(f, delay, interval));
     }
-
-    public static void ExitGame(this MonoBehaviour mb) {
-#if UNITY_EDITOR
-      UnityEditor.EditorApplication.isPlaying = false;
-#else
-      UnityEngine.Application.Quit();
-#endif
-    }
   }
 }