+++ Runtime/Utils/UniStart.cs
@@ -2,17 +2,26 @@ using System;
 using UnityEngine.Events;
 
 namespace DT.UniStart {
-  public static class FnHelper {
+  public static class UniStart {
+    public static void ExitGame() {
+#if UNITY_EDITOR
+      UnityEditor.EditorApplication.isPlaying = false;
+#else
+      UnityEngine.Application.Quit();
+#endif
+    }
+
+    #region Fn Helpers
     public static UnityAction Fn(UnityAction action) => action;
     public static UnityAction<T0> Fn<T0>(UnityAction<T0> action) => action;
     public static UnityAction<T0, T1> Fn<T0, T1>(UnityAction<T0, T1> action) => action;
     public static UnityAction<T0, T1, T2> Fn<T0, T1, T2>(UnityAction<T0, T1, T2> action) => action;
     public static UnityAction<T0, T1, T2, T3> Fn<T0, T1, T2, T3>(UnityAction<T0, T1, T2, T3> action) => action;
-
     public static Func<R> Fn<R>(Func<R> f) => f;
     public static Func<T0, R> Fn<T0, R>(Func<T0, R> f) => f;
     public static Func<T0, T1, R> Fn<T0, T1, R>(Func<T0, T1, R> f) => f;
     public static Func<T0, T1, T2, R> Fn<T0, T1, T2, R>(Func<T0, T1, T2, R> f) => f;
     public static Func<T0, T1, T2, T3, R> Fn<T0, T1, T2, T3, R>(Func<T0, T1, T2, T3, R> f) => f;
+    #endregion
   }
 }